fix: scriptlets missing brackets
