+++ .changeset/red-rabbits-type.md
@@ -1,0 +1,5 @@
+---
+"prettier-plugin-marko": patch
+---
+
+Fix issue with scriptlets being considered safe to avoid wrapping with brackets even when it was not.
