+++ src/__tests__/fixtures/scriptlet-long-enclose/template.marko
@@ -1,0 +1,1 @@
+$ const isWebpSupported = $global.request.headers.get("accept")?.includes("image/webp");
